第一次

+++ design_manager/src/main/java/com/sakura/pojo/Stupaperdetails.java
@@ -21,7 +21,7 @@ public class Stupaperdetails implements Serializable {
 
     private Integer state;
 
-    private Integer score;
+    private Double score;
 
     private Long pid;
 
@@ -99,11 +99,11 @@ public class Stupaperdetails implements Serializable {
         this.state = state;
     }
 
-    public Integer getScore() {
+    public Double getScore() {
         return score;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(Double score) {
         this.score = score;
     }
